refactor: replace config.py with config.example.py
This is based on industry best practices and standards. config.py should
be treated and used the same as environment variables, holding the users
private information. The config.example.py is better becuase when you
make a new release and have to change config.py, the developers would
have to rewrite all of their information again but now all they need to
do is copy paste it. This also solves issues with git and tracking and
reduces the potential risk of leaking your private personal information.